convert cmd/test to use testing package
